backend/requirements.txt
fastapi
uvicorn
python-multipart
scikit-learn
nltk
PyPDF2
python-dotenv
groq
